Move boundedScan utilities to new module
