+++ lib/ChainwebDb/Queries.hs
@@ -1,3 +1,8 @@
+-- We're disabling the missing-signatures warning because some Beam types are
+-- too unwieldy to type and some types aren't even exported so they can't even
+-- be typed explicitly.
+{-# OPTIONS_GHC -Wno-missing-signatures #-}
+
 {-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE ExistentialQuantification #-}
@@ -9,7 +14,6 @@
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE StandaloneDeriving #-}
-{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 -- |
 
@@ -23,7 +27,6 @@ import           Data.Maybe (maybeToList)
 import           Data.Text (Text)
 import           Data.Time
 import           Database.Beam hiding (insert)
-import           Database.Beam.Query.Internal (QOrd, QNested)
 import           Database.Beam.Postgres
 import           Database.Beam.Postgres.Syntax
 import           Database.Beam.Backend.SQL.SQL92
@@ -33,6 +36,7 @@ import           Chainweb.Api.ChainId
 import           Chainweb.Api.Common (BlockHeight)
 import           ChainwebData.Api
 import           ChainwebData.Pagination
+import           ChainwebDb.BoundedScan
 import           ChainwebDb.Database
 import           ChainwebDb.Types.Block
 import           ChainwebDb.Types.DbHash
@@ -120,12 +124,6 @@ data EventCursorT f = EventCursor
 type EventCursor = EventCursorT Identity
 deriving instance Show EventCursor
 
-eventsCursorOrder ::
-  EventT (PgExpr s) ->
-  ( QOrd Postgres s Int64
-  , QOrd Postgres s ReqKeyOrCoinbase
-  , QOrd Postgres s Int64
-  )
 eventsCursorOrder ev =
   ( desc_ $ _ev_height ev
   , desc_ $ _ev_requestkey ev
@@ -170,49 +168,6 @@ eventsSearchOffset ::
 eventsSearchOffset esp eqs =
   boundedScanOffset (eventSearchCond esp) (eventsAfterStart eqs) eventsCursorOrder eventToCursor
 
-type QBS = QBaseScope
-type NQBS = QNested QBS
-type N2QBS = QNested NQBS
-type N3QBS = QNested N2QBS
-type N4QBS = QNested N3QBS
-
-boundedScanOffset :: forall a db rowT cursorT.
-  (SqlOrderable Postgres a, Beamable rowT, Beamable cursorT) =>
-  (forall s. rowT (PgExpr s) -> PgExpr s Bool) ->
-  Q Postgres db N3QBS (rowT (PgExpr N3QBS)) ->
---  (forall s. Q Postgres ChainwebDataDb s (EventT (PgExpr s))) ->
-  (rowT (PgExpr N3QBS) -> a) ->
-  (rowT (PgExpr N3QBS) -> cursorT (PgExpr N3QBS)) ->
-  Offset ->
-  Int64 ->
-  SqlSelect Postgres (cursorT Identity, Int64, Int64)
-boundedScanOffset condExp toScan order toCursor (Offset o) scanLimit =
-  select $ limit_ 1 noLimitQuery where
-  -- For some reason, beam fails to unify the types here unless we move noLimitQuery
-  -- to a separate definition and explicitly specify its type
-  noLimitQuery :: Q Postgres db NQBS
-    ( cursorT (PgExpr NQBS)
-    , PgExpr NQBS Int64
-    , PgExpr NQBS Int64
-    )
-  noLimitQuery = do
-    (cursor, matchingRow, scan_num, found_num) <- subselect_ $ withWindow_
-      (\ev ->
-        ( frame_ (noPartition_ @Int) (Just $ order ev) noBounds_
-        , frame_ (noPartition_ @Int) (Just $ order ev) (fromBound_ unbounded_)
-        )
-      )
-      (\ev (wNoBounds, wTrailing) ->
-        ( toCursor ev
-        , condExp ev
-        , rowNumber_ `over_` wNoBounds
-        , countAll_ `filterWhere_` condExp ev `over_` wTrailing
-        )
-      )
-      toScan
-    guard_ $ scan_num ==. val_ scanLimit ||. (matchingRow &&. found_num ==. val_ (fromInteger o))
-    return (cursor, found_num, scan_num)
-
 eventsSearchLimit ::
   EventSearchParams ->
   EventQueryStart ->
@@ -225,28 +180,6 @@ eventsSearchLimit esp eqs limit scanLimit = select $ do
   blk <- all_ $ _cddb_blocks database
   guard_ $ _ev_block ev `references_` blk
   return (ev, blk, scan_num, matchingRow)
-
-boundedScanLimit ::
-  (SqlOrderable Postgres a, Beamable rowT) =>
-  (rowT (PgExpr NQBS) -> (PgExpr NQBS) Bool) ->
-  Q Postgres db N4QBS (rowT (PgExpr N4QBS)) ->
-  (rowT (PgExpr N4QBS) -> a) ->
-  Limit ->
-  Int64 ->
-  Q Postgres db QBS (rowT (PgExpr QBS), PgExpr QBS Int64, PgExpr QBS Bool)
-boundedScanLimit cond toScan order (Limit l) scanLimit = limit_ l $ do
-  (ev, scan_num) <- subselect_ $ limit_ (fromIntegral scanLimit) $ withWindow_
-    (\ev -> frame_ (noPartition_ @Int) (Just $ order ev) noBounds_)
-    (\ev window ->
-      ( ev
-      , rowNumber_ `over_` window
-      )
-    )
-    toScan
-  let scan_end = scan_num ==. val_ scanLimit
-      matchingRow = cond ev
-  guard_ $ scan_end ||. matchingRow
-  return (ev, scan_num, matchingRow)
 
 _bytequery :: Sql92SelectSyntax (BeamSqlBackendSyntax be) ~ PgSelectSyntax => SqlSelect be a -> ByteString
 _bytequery = \case
